Sett valg av oppgavestatuser i egen komponent
